+++ Hurrican/src/bosses/Boss_RiesenRaupe.cpp
@@ -57,11 +57,11 @@ void GegnerRiesenRaupe::DoKI() {
                 AnimEnde = 13;
                 AnimSpeed = 1.5f;
 
-                PartikelSystem.PushPartikel(xPos + 30, yPos - 2, SMOKE);
-                PartikelSystem.PushPartikel(xPos + 120, yPos - 2, SMOKE);
+                PartikelSystem.PushPartikel(xPos + 30.0f, yPos - 2.0f, SMOKE);
+                PartikelSystem.PushPartikel(xPos + 120.0f, yPos - 2.0f, SMOKE);
 
-                Projectiles.PushProjectile(xPos + 30, yPos - 2, PFLANZESHOT, pAim);
-                Projectiles.PushProjectile(xPos + 120, yPos - 2, PFLANZESHOT, pAim);
+                Projectiles.PushProjectile(xPos + 30.0f, yPos - 2.0f, PFLANZESHOT, pAim);
+                Projectiles.PushProjectile(xPos + 120.0f, yPos - 2.0f, PFLANZESHOT, pAim);
 
                 SoundManager.PlayWave(100, 128, 11025, SOUND_ROCKET);
             }
@@ -108,15 +108,19 @@ void GegnerRiesenRaupe::DoKI() {
 void GegnerRiesenRaupe::GegnerExplode() {
     // Explosionen erzeugen
     for (int i = 0; i < 10; i++) {
-        PartikelSystem.PushPartikel(xPos + random(180) - 30, yPos + random(60) - 30, EXPLOSION_BIG);
-        PartikelSystem.PushPartikel(xPos + random(180) - 30, yPos + random(60) - 30, SPIDERSPLITTER);
+        PartikelSystem.PushPartikel(xPos + static_cast<float>(random(180) - 30),
+                                    yPos + static_cast<float>(random(60) - 30), EXPLOSION_BIG);
+        PartikelSystem.PushPartikel(xPos + static_cast<float>(random(180) - 30),
+                                    yPos + static_cast<float>(random(60) - 30), SPIDERSPLITTER);
     }
 
     for (int i = 0; i < 20; i++)
-        PartikelSystem.PushPartikel(xPos + random(180) - 10, yPos + random(60) - 10, EXPLOSION_MEDIUM2);
+        PartikelSystem.PushPartikel(xPos + static_cast<float>(random(180) - 10),
+                                    yPos + static_cast<float>(random(60) - 10), EXPLOSION_MEDIUM2);
 
     for (int i = 0; i < 300; i++)
-        PartikelSystem.PushPartikel(xPos + random(180) - 10, yPos + random(50), ROCKETSMOKE);
+        PartikelSystem.PushPartikel(xPos + static_cast<float>(random(180) - 10),
+                                    yPos + static_cast<float>(random(50)), ROCKETSMOKE);
 
     SoundManager.PlayWave(100, 128, 11025 + random(2000), SOUND_EXPLOSION3);  // Sound ausgeben
